【fix】rename plugin name for mq-consume-deny
