+++ sermant-agentcore/sermant-agentcore-config/config/all/plugins.yaml
@@ -7,6 +7,7 @@ plugins:
   - threadlocal
   - service-registry
   - loadbalancer
+  - mq-consume-deny
 adaptors:
   - lubanops
 profiles:
