--- sermant-plugins/sermant-message/pom.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>sermant-plugins</artifactId>
-        <groupId>com.huaweicloud.sermant</groupId>
-        <version>1.0.0</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>sermant-message</artifactId>
-    <name>sermant-message</name>
-    <modules>
-        <module>kafka-consumer-plugin</module>
-        <module>rabbitmq-consumer-plugin</module>
-        <module>message-common</module>
-    </modules>
-    <packaging>pom</packaging>
-
-    <properties>
-        <sermant.basedir>${pom.basedir}/../../..</sermant.basedir>
-        <package.plugin.name>deny-consume</package.plugin.name>
-    </properties>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>com.huaweicloud.sermant</groupId>
-                <artifactId>message-common</artifactId>
-                <version>${project.version}</version>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <profiles>
-        <profile>
-            <id>agent</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <modules>
-                <module>kafka-consumer-plugin</module>
-                <module>rabbitmq-consumer-plugin</module>
-                <module>message-common</module>
-            </modules>
-        </profile>
-        <profile>
-            <id>all</id>
-            <modules>
-                <module>kafka-consumer-plugin</module>
-                <module>rabbitmq-consumer-plugin</module>
-                <module>message-common</module>
-            </modules>
-        </profile>
-        <profile>
-            <id>release</id>
-            <modules>
-                <module>kafka-consumer-plugin</module>
-                <module>rabbitmq-consumer-plugin</module>
-                <module>message-common</module>
-            </modules>
-        </profile>
-    </profiles>
-</project>
