+++ sermant-plugins/sermant-mq-consume-deny/kafka-consumer-plugin/src/main/java/com/huaweicloud/sermant/kafka/mock/MockKafkaConsumer.java
