+++ sermant-plugins/sermant-mq-consume-deny/kafka-consumer-plugin/src/main/resources/META-INF/services/com.huaweicloud.sermant.core.plugin.agent.declarer.PluginDeclarer
