+++ sermant-plugins/sermant-mq-consume-deny/message-common/src/main/java/com/huaweicloud/sermant/message/common/utils/MockUtils.java
