+++ sermant-plugins/sermant-mq-consume-deny/message-common/src/main/resources/META-INF/services/com.huaweicloud.sermant.core.plugin.config.PluginConfig
