+++ sermant-plugins/sermant-mq-consume-deny/rabbitmq-consumer-plugin/src/main/resources/META-INF/services/com.huaweicloud.sermant.core.plugin.agent.declarer.PluginDeclarer
